Rename testErrorCodes to testEnums
