+++ simplex/test.h
@@ -1,16 +1,6 @@
 #pragma once
 
-#include "testErrorCodes.h"
-
-enum testList {
-	SIMPL_TEST = 0,
-	NEG_SIZE_TEST,
-	NEGROW_TEST,
-	UNBOUNDED_TEST,
-	INFEASIBLE_TEST,
-
-	NUMBER_OF_TESTS
-};
+#include "testEnums.h"
 
 class allTests {
 public:
