+++ simplex/testEnums.h
@@ -2,6 +2,16 @@
 
 #include <utility>
 
+enum testList {
+	SIMPL_TEST = 0,
+	NEG_SIZE_TEST,
+	NEGROW_TEST,
+	UNBOUNDED_TEST,
+	INFEASIBLE_TEST,
+
+	NUMBER_OF_TESTS
+};
+
 enum testErrorCodes {
 	TEST_SUCCESS,
 	TEST_INCORRECT_VALUE,
